// public/server/src/typechain/CheeseChain.ts
export declare namespace CheeseChain {
  export type ParticipantStruct = {
    name: PromiseOrValue<string>;
    role: PromiseOrValue<BigNumberish>;
    owner: PromiseOrValue<string>;
  };

  export type ParticipantStructOutput = [string, number, string] & {
    name: string;
    role: number;
    owner: string;
  };

  export type CoordinatesStruct = {
    latitude: PromiseOrValue<string>;
    longitude: PromiseOrValue<string>;
  };

  export type CoordinatesStructOutput = [string, string] & {
    latitude: string;
    longitude: string;
  };

  export type TestResultStruct = {
    result: PromiseOrValue<boolean>;
    timestamp: PromiseOrValue<BigNumberish>;
  };

  export type TestResultStructOutput = [boolean, BigNumber] & {
    result: boolean;
    timestamp: BigNumber;
  };

  export type LotStruct = {
    testResult: CheeseChain.TestResultStruct;
    lastStep: PromiseOrValue<BigNumberish>;
    timestamp: PromiseOrValue<BigNumberish>;
    milkBatchId: PromiseOrValue<BigNumberish>[];
  };

  export type LotStructOutput = [
    CheeseChain.TestResultStructOutput,
    BigNumber,
    BigNumber,
    BigNumber[]
  ] & {
    testResult: CheeseChain.TestResultStructOutput;
    lastStep: BigNumber;
    timestamp: BigNumber;
    milkBatchId: BigNumber[];
  };
}